ai-headhunter-assistant-prod/apps/backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF 解析
pdfplumber==0.10.3
Pillow==10.1.0

# OpenAI（兼容通义千问）
openai==2.7.2

# 工具
python-dotenv==1.0.0

# 飞书 SDK
lark-oapi>=1.2.0

